+++ tests/globals,toolarge.R
@@ -3,25 +3,26 @@ library("listenv")
 
 message("*** Globals - too large ...")
 
-ooptsT <- options(future.globals.maxSize = object.size(1:1000) - 1L)
+a <- integer(length = 1000)
+ooptsT <- options(future.globals.maxSize = object.size(a) - 100L)
 limit <- getOption("future.globals.maxSize")
 cat(sprintf("Max total size of globals: %g bytes\n", limit))
 
 plan(multisession)
 
 exprs <- list(
-  A = substitute({ a }, env = list()),
-  B = substitute({ a * b }, env = list()),
-  C = substitute({ a * b * c }, env = list()),
-  D = substitute({ a * b * c * d }, env = list()),
+  A = substitute({ a                 }, env = list()),
+  B = substitute({ a * b             }, env = list()),
+  C = substitute({ a * b * c         }, env = list()),
+  D = substitute({ a * b * c * d     }, env = list()),
   E = substitute({ a * b * c * d * e }, env = list())
 )
 
-a <- 1:1000
-b <- 1:900
-c <- 1:800
-d <- 1:700
-e <- 1
+a <- integer(length = 1000)
+b <- integer(length =  900)
+c <- integer(length =  800)
+d <- integer(length =  700)
+e <- integer(length =    1)
 
 for (name in names(exprs)) {
   message(sprintf("Expression %s:", name))